requirements.txt
python-dotenv==0.13.0
python-telegram-bot==12.8
requests==2.24.0
